legacy-python-gui-plot/requirements.txt
SumoNetVis==1.6.0
matplotlib==3.5.2
imageio
seaborn
numpy
progressbar2
PyYAML